backend/requirements.txt
fastapi
uvicorn
langchain
langchain-ollama
langchain-community
duckduckgo-search
wikipedia
pydantic
python-dotenv
ddgs
